refactor(procedure): extract the email templates concern
Procedure gathers the six email template associations, their readers and
the attestation consistency check. Pure move, ahead of the declarative
variants that grow all three.
